Face_Recognition/requirements.txt
deepface
faiss-cpu
opencv-python 
json
